Add ability to get lexer diagnostics
When parsing tokens only, we currently have no way to get access to the
lexer's diagnostics. This adds an optional out parameter.

The real fix, of course, would be to mirror what Roslyn is doing, which is
exposing the diagnostics on the actual tokens, which makes the output neatly
self-contained.