Adding parsing for local functions
Function declarations now need to derive from StatementSyntax so that we can use them anywhere we can use statements. MemberSyntax is no longer needed - it was necessary to make functions usable as statements in a special scenario, but now they're always statements. I don't think we need to have separate MethodDeclarationSyntax (which would be a MemberSyntax) and LocalFunctionStatementSyntax like C# does, at least for now, because we don't even have classes and don't need to differentiate these.